Finishing more power ups

+++ Assets/PowerUpData.cs
@@ -7,5 +7,5 @@ public class PowerUpData : ScriptableObject {
     public string itemName;
     public Sprite overlaySprite;
     public GameObject prefab;
-    public int dropRate;
+    public float dropRate;
 }
